+++ testlib/src/test/java/com/diffplug/spotless/biome/BiomeStepTest.java
@@ -15,8 +15,11 @@
  */
 package com.diffplug.spotless.biome;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import org.junit.jupiter.api.BeforeAll;
@@ -53,7 +56,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCjs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
 		}
@@ -64,7 +67,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCts() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.cts", "biome/ts/fileAfter.cts");
 		}
@@ -76,7 +79,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testAutoDetectCssExperimental() {
 			var path = createBiomeConfig("biome/config/css-enabled.json");
-			var step = BiomeStep.withExeDownload("1.8.3", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.8.3", downloadDir).withConfigPath(path.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -87,7 +90,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCssStable() {
-			var step = BiomeStep.withExeDownload("1.9.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.9.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -98,7 +101,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.js", "biome/js/fileAfter.js");
 		}
@@ -109,7 +112,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJson() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/fileBefore.json", "biome/json/fileAfter.json");
 		}
@@ -120,7 +123,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJsonc() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/jsonc/fileBefore.jsonc", "biome/jsonc/fileAfter.jsonc");
 		}
@@ -131,7 +134,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJsx() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.jsx", "biome/js/fileAfter.jsx");
 		}
@@ -142,7 +145,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectMjs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.mjs", "biome/js/fileAfter.mjs");
 		}
@@ -153,7 +156,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectMts() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.mts", "biome/ts/fileAfter.mts");
 		}
@@ -164,7 +167,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectTs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.ts", "biome/ts/fileAfter.ts");
 		}
@@ -175,7 +178,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectTsx() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.tsx", "biome/ts/fileAfter.tsx");
 		}
@@ -188,7 +191,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void preservesIgnoredFiles() {
-			var step = BiomeStep.withExeDownload("1.5.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.5.0", downloadDir).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/package.json", "biome/json/packageAfter.json");
 		}
@@ -228,7 +231,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testLineWidth120() {
 			var path = createBiomeConfig("biome/config/line-width-120.json");
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withConfigPath(path.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter120.js");
 		}
@@ -239,11 +242,36 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testLineWidth80() {
 			var path = createBiomeConfig("biome/config/line-width-80.json");
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withConfigPath(path.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter80.js");
 		}
 
+		/**
+		 * Test the path to the JSON config file can be specified directly (requires biome 2.x).
+		 */
+		@Test
+		void testPathToFile() {
+			var path = createBiomeConfig("biome/config/line-width-80.json").resolve("biome.json");
+			assertTrue(Files.isRegularFile(path));
+			assertTrue(Files.exists(path));
+			var step = BiomeStep.withExeDownload("2.1.1", downloadDir).withConfigPath(path.toString()).create();
+			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
+			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter80.js");
+		}
+
+		/**
+		 * Test the path to the folder with the biome.json config file can be specified.
+		 */
+		@Test
+		void testPathToDirectory() {
+			var path = createBiomeConfig("biome/config/line-width-80.json");
+			assertTrue(Files.isDirectory(path));
+			assertTrue(Files.exists(path));
+			var step = BiomeStep.withExeDownload("2.1.1", downloadDir).withConfigPath(path.toString()).create();
+			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
+			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter80.js");
+		}
 	}
 
 	/**
@@ -257,7 +285,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCjs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
 		}
@@ -269,7 +297,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testSetLanguageCssExperimental() {
 			var path = createBiomeConfig("biome/config/css-enabled.json");
-			var step = BiomeStep.withExeDownload("1.8.3", downloadDir.toString()).withConfigPath(path).withLanguage("css").create();
+			var step = BiomeStep.withExeDownload("1.8.3", downloadDir).withConfigPath(path.toString()).withLanguage("css").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -280,7 +308,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCssStable() {
-			var step = BiomeStep.withExeDownload("1.9.0", downloadDir.toString()).withLanguage("css").create();
+			var step = BiomeStep.withExeDownload("1.9.0", downloadDir).withLanguage("css").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -291,7 +319,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCts() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.cts", "biome/ts/fileAfter.cts");
 		}
@@ -302,7 +330,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.js", "biome/js/fileAfter.js");
 		}
@@ -313,7 +341,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJson() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("json").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("json").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/fileBefore.json", "biome/json/fileAfter.json");
 		}
@@ -324,7 +352,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJsonc() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("jsonc").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("jsonc").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/jsonc/fileBefore.jsonc", "biome/jsonc/fileAfter.jsonc");
 		}
@@ -335,7 +363,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJsx() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("jsx").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("jsx").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.jsx", "biome/js/fileAfter.jsx");
 		}
@@ -346,7 +374,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageMjs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.mjs", "biome/js/fileAfter.mjs");
 		}
@@ -357,7 +385,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageMts() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.mts", "biome/ts/fileAfter.mts");
 		}
@@ -368,7 +396,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageTs() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.ts", "biome/ts/fileAfter.ts");
 		}
@@ -379,17 +407,17 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageTsx() {
-			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("tsx").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).withLanguage("tsx").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.tsx", "biome/ts/fileAfter.tsx");
 		}
 	}
 
-	private String createBiomeConfig(String name) {
+	private Path createBiomeConfig(String name) {
 		var config = createTestFile(name).toPath();
 		var dir = config.getParent();
-		var rome = dir.resolve("biome.json");
-		ThrowingEx.run(() -> Files.copy(config, rome));
-		return dir.toString();
+		var biomeConfig = dir.resolve("biome.json");
+		ThrowingEx.run(() -> Files.copy(config, biomeConfig));
+		return dir;
 	}
 }